+++ examples/tv-casting-app/tv-casting-common/commands/clusters/ReportCommand.h
@@ -105,7 +105,7 @@ public:
         SetCommandExitStatus(mError);
     }
 
-    void OnSubscriptionEstablished(uint64_t subscriptionId) override { OnAttributeSubscription(); }
+    void OnSubscriptionEstablished(uint32_t subscriptionId) override { OnAttributeSubscription(); }
 
 protected:
     CHIP_ERROR ReportAttribute(ChipDevice * device, std::vector<chip::EndpointId> endpointIds,
