+++ src/app/BufferedReadCallback.h
@@ -76,7 +76,7 @@ private:
     }
 
     void OnDone() override { return mCallback.OnDone(); }
-    void OnSubscriptionEstablished(uint64_t aSubscriptionId) override { mCallback.OnSubscriptionEstablished(aSubscriptionId); }
+    void OnSubscriptionEstablished(uint32_t aSubscriptionId) override { mCallback.OnSubscriptionEstablished(aSubscriptionId); }
 
     void OnDeallocatePaths(chip::app::ReadPrepareParams && aReadPrepareParams) override
     {
