+++ src/app/ClusterStateCache.h
@@ -569,7 +569,7 @@ private:
         return mCallback.OnDone();
     }
 
-    void OnSubscriptionEstablished(uint64_t aSubscriptionId) override { mCallback.OnSubscriptionEstablished(aSubscriptionId); }
+    void OnSubscriptionEstablished(uint32_t aSubscriptionId) override { mCallback.OnSubscriptionEstablished(aSubscriptionId); }
 
     void OnDeallocatePaths(chip::app::ReadPrepareParams && aReadPrepareParams) override
     {
