+++ src/app/InteractionModelEngine.h
@@ -119,7 +119,7 @@ public:
      * @retval #CHIP_ERROR_KEY_NOT_FOUND If the subscription is not found.
      * @retval #CHIP_NO_ERROR On success.
      */
-    CHIP_ERROR ShutdownSubscription(uint64_t aSubscriptionId);
+    CHIP_ERROR ShutdownSubscription(uint32_t aSubscriptionId);
 
     /**
      * Tears down active subscriptions for a given peer node ID.
@@ -259,7 +259,10 @@ public:
     //
     // Get direct access to the underlying read handler pool
     //
-    auto & GetReadHandlerPool() { return mReadHandlers; }
+    auto & GetReadHandlerPool()
+    {
+        return mReadHandlers;
+    }
 
     //
     // Override the maximal capacity of the underlying read handler pool to mimic
@@ -270,7 +273,10 @@ public:
     //
     // If -1 is passed in, no override is instituted and default behavior resumes.
     //
-    void SetHandlerCapacity(int32_t sz) { mReadHandlerCapacityOverride = sz; }
+    void SetHandlerCapacity(int32_t sz)
+    {
+        mReadHandlerCapacityOverride = sz;
+    }
 
     //
     // Override the maximal capacity of the underlying attribute path pool and event path pool to mimic
@@ -281,7 +287,10 @@ public:
     //
     // If -1 is passed in, no override is instituted and default behavior resumes.
     //
-    void SetPathPoolCapacity(int32_t sz) { mPathPoolCapacityOverride = sz; }
+    void SetPathPoolCapacity(int32_t sz)
+    {
+        mPathPoolCapacityOverride = sz;
+    }
 
     //
     // Override the maximal capacity of the underlying read handler pool to mimic
@@ -310,7 +319,10 @@ public:
     // enforce such check based on the configured size. This flag is used for unit tests only, there is another compare time flag
     // CHIP_CONFIG_IM_FORCE_FABRIC_QUOTA_CHECK for stress tests.
     //
-    void SetForceHandlerQuota(bool forceHandlerQuota) { mForceHandlerQuota = forceHandlerQuota; }
+    void SetForceHandlerQuota(bool forceHandlerQuota)
+    {
+        mForceHandlerQuota = forceHandlerQuota;
+    }
 
     //
     // When testing subscriptions using the high-level APIs in src/controller/ReadInteraction.h,
@@ -346,7 +358,10 @@ private:
     void OnDone(CommandHandler & apCommandObj) override;
     void OnDone(ReadHandler & apReadObj) override;
 
-    ReadHandler::ApplicationCallback * GetAppCallback() override { return mpReadHandlerApplicationCallback; }
+    ReadHandler::ApplicationCallback * GetAppCallback() override
+    {
+        return mpReadHandlerApplicationCallback;
+    }
 
     CHIP_ERROR OnUnsolicitedMessageReceived(const PayloadHeader & payloadHeader, ExchangeDelegate *& newDelegate) override;
